pulling_lyrics.py
a python-workflow that parses through the wikipedia pages to find the top 100 Billboard songs 1956-2014. Once artist and song name are extracted, the song is looled up at lyricsmode.  The corresponding website is parsed for the text and then written into a file (one json file for each year including 100 lyrics). Not found lyrics are ignored.